+++ terraform/variables.tf
@@ -587,6 +587,18 @@ variable "monitoring_account" {
   description = "Suffix to add to the alarm name to distinguish the account sending the warning"
 }
 
+variable "alarm_evaluation_periods" {
+  type        = number
+  default     = 10
+  description = "Number of consecutive periods the metric must breach the threshold before the alarm triggers (default 10 x 60s = 10 minutes)"
+}
+
+variable "alarm_period_seconds" {
+  type        = number
+  default     = 60
+  description = "Length of each evaluation period in seconds for ALB health alarms"
+}
+
 ###############
 # DATA MIGRATION
 ###############